scripts/flippigator/requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==22.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
iniconfig==2.0.0
isort==5.10.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pathspec==0.11.0
Pillow==9.5.0
platformdirs==3.1.1
pluggy==1.0.0
protobuf==4.23.1
pyserial==3.5
pyserial-asyncio==0.6
pytest==7.2.0
pytest-asyncio==0.23.6
six==1.16.0
termcolor==2.1.0
tomli==2.0.1
